+++ HISTORY.md
@@ -10,6 +10,7 @@
 * New Romanian `ro` locale.
 * Added `f.Finance.RoutingNumber` - Generates an ABA routing number with valid check digit.
 * Added `Faker.GenerateForever` that returns `IEnumerable<T>` with unlimited generated items when iterated over.
+* Added United Kingdom extension method to generate bank ShortCodes on `f.Finance.ShortCode()`.
 
 ## v18.0.2
 * Issue 86: Removed diacritic mark/accents (á, í, ó, ú, etc) from generated email addresses and user names.